Move get_active_validators to utils
